division complete

+++ division.rb
@@ -4,4 +4,4 @@ val1 = gets;
 print "Please enter number 2 : "
 val2 = gets;
 #change val to integer (i)
-print "Answer : " , (val1.to_i / val2.to_i), "\n"
+print "Answer : " , (val1.to_f / val2.to_f), "\n"
